+++ example/Menus.jsx
@@ -1,13 +1,19 @@
 import React from 'react'
 import Menu from '../components/Menu'
 import AnimatedMenu from '../components/AnimatedMenu'
-import { PrimaryTitle, Subtitle, Paragraph } from '../components/Text'
+import Label from '../components/Label'
+import { SecondaryTitle, Subtitle, Paragraph } from '../components/Text'
 import Code from './Code'
 
 const options = [
   { key: 'home', label: 'Home' },
   { key: 'faq', label: 'FAQ' },
   { key: 'archive', label: 'Archive' }
+]
+
+const optionsWithComponents = [
+  { key: 'home', label: <div>Home</div> },
+  { key: 'faq', label: <div>To do <Label design='warning'>urgent</Label></div> }
 ]
 
 class AddableMenu extends React.Component {
@@ -61,10 +67,10 @@ class AddableMenu extends React.Component {
 export default function Menus () {
   return (
     <div>
-      <PrimaryTitle small blue>Tabs</PrimaryTitle>
+      <SecondaryTitle margins small color='blue'>Tabs</SecondaryTitle>
 
-      <Subtitle>Fluid</Subtitle>
-      <Paragraph>
+      <Subtitle margins>Fluid</Subtitle>
+      <Paragraph margins>
         Click and check your console.
       </Paragraph>
       <Code>
@@ -76,8 +82,8 @@ export default function Menus () {
         />
       </Code>
 
-      <Subtitle>Static</Subtitle>
-      <Paragraph>
+      <Subtitle margins>Static</Subtitle>
+      <Paragraph margins>
         Click and check your console.
       </Paragraph>
       <Code>
@@ -89,12 +95,12 @@ export default function Menus () {
           options={options} />
       </Code>
 
-      <Subtitle>White</Subtitle>
-      <Paragraph>
+      <Subtitle margins>White</Subtitle>
+      <Paragraph margins>
         Click and check your console.
       </Paragraph>
       <Code>
-        <div style={{background: '#333', padding: '20px'}}>
+        <div style={{background: '#0074c8', padding: '20px'}}>
           <Menu
             white
             onChange={(key) => console.log('You selected', key)}
@@ -105,35 +111,35 @@ export default function Menus () {
         </div>
       </Code>
 
-      <PrimaryTitle small blue>Animated</PrimaryTitle>
+      <SecondaryTitle margins small color='blue'>Animated</SecondaryTitle>
 
-      <Paragraph>
+      <Paragraph margins>
         Animated menus must be controlled. On implementation reference can be found in examples/Menus.jsx.
       </Paragraph>
 
-      <Subtitle>Fluid</Subtitle>
-      <Paragraph>Click "Add" and check it out</Paragraph>
+      <Subtitle margins>Fluid</Subtitle>
+      <Paragraph margins>Click "Add" and check it out</Paragraph>
       <Code>
         <AddableMenu name='addable-fluid' />
       </Code>
 
-      <Paragraph>Click "Add" and check it out</Paragraph>
-      <Subtitle>Static</Subtitle>
+      <Paragraph margins>Click "Add" and check it out</Paragraph>
+      <Subtitle margins>Static</Subtitle>
       <Code>
         <AddableMenu name='addable-static' tabDisplay='static' />
       </Code>
 
-      <Subtitle>White</Subtitle>
+      <Subtitle margins>White</Subtitle>
       <Code>
-        <div style={{background: '#333', padding: '20px'}}>
+        <div style={{background: '#0074c8', padding: '20px'}}>
           <AddableMenu white name='addable-white' tabDisplay='static' />
         </div>
       </Code>
 
-      <PrimaryTitle small blue>Segmented</PrimaryTitle>
+      <SecondaryTitle margins small color='blue'>Segmented</SecondaryTitle>
 
-      <Subtitle>Fluid</Subtitle>
-      <Paragraph>
+      <Subtitle margins>Fluid</Subtitle>
+      <Paragraph margins>
         Click and check your console.
       </Paragraph>
       <Code>
@@ -146,8 +152,8 @@ export default function Menus () {
         />
       </Code>
 
-      <Subtitle>Static</Subtitle>
-      <Paragraph>
+      <Subtitle margins>Static</Subtitle>
+      <Paragraph margins>
         Click and check your console.
       </Paragraph>
       <Code>
@@ -160,6 +166,18 @@ export default function Menus () {
           options={options} />
       </Code>
 
+      <Subtitle margins>Options with components</Subtitle>
+      <Paragraph margins>
+        Options also support components as labels.
+      </Paragraph>
+      <Code>
+        <Menu
+          onChange={(key) => console.log('You selected', key)}
+          tabDisplay='static'
+          name='options-with-components'
+          selected='home'
+          options={optionsWithComponents} />
+      </Code>
     </div>
   )
 }